requirements.txt
aniso8601==9.0.1
APScheduler==3.7.0
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
click==8.0.4
Flask==1.1.2
flask-restplus==0.13.0
idna==2.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
pyrsistent==0.18.0
pytz==2021.3
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
typing_extensions==4.1.1
tzlocal==2.1
urllib3==1.26.6
Werkzeug==0.16.1
zipp==3.6.0
